+++ src/shared/ui/DialogHeader.tsx
@@ -1,0 +1,8 @@
+type DialogHeaderProps = React.HTMLAttributes<HTMLDivElement> & {
+  className?: string;
+};
+
+export const DialogHeader = ({ className = '', ...props }: DialogHeaderProps) => (
+  <div className={`flex flex-col space-y-1.5 text-center sm:text-left ${className}`} {...props} />
+);
+DialogHeader.displayName = 'DialogHeader';
